Regal patches for 8dda43495320536adcfcf0b50e32c763d3b603d7

+++ common/trace_dump.cpp
@@ -235,7 +235,7 @@ public:
     }
 
     void visit(Backtrace & backtrace) {
-        for (int i = 0; i < backtrace.size(); i ++) {
+        for (size_t i = 0; i < backtrace.size(); i ++) {
             visit(backtrace[i]);
             os << "\n";
         }
